requirements.txt
pytest-playwright==0.4.4
pydantic==2.5.3
tqdm==4.66.1
beautifulsoup4==4.12.2
hydra-core==1.3.2
pandas==2.2.0
nltk==3.8.1